final code

+++ src/main/java/br/com/fsantosinfo/compositekeyidclass/repository/ErrorGameRepository.java
@@ -16,10 +16,9 @@ public interface ErrorGameRepository extends JpaRepository<ErrorGame, ErrorGameP
     @Query(value = "SELECT NVL(MAX(ERROR_SEQUENCE_NUMBER), 0) + 1 maxSequence "
     + " FROM TB_ERROR_GAME "
     + "WHERE GAME_ID = :gameId AND GAME_PLATFORM = :gamePlatform ", nativeQuery = true)
-    Integer getNextErrorSequenceCodeValue(
-        @Param("gameId") Long gameId,
-        @Param("gamePlatform") String gamePlatform);
-    
+    Long getNextErrorSequenceCodeValue(@Param("gameId") Long gameId, @Param("gamePlatform") String gamePlatform);
+    // we're using this method when we don't want to use Oracle Sequences. But it works only with IdClass and not embedded 
+
     Page<ErrorGame> findByGameIdAndGamePlatform(Long gameId, String gamePlatform, Pageable pageable);
     // Method reference (using the attribute name) only works in composite key with @IdClass
 }
